fastapi/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-dotenv==1.0.1
langchain==0.1.9
langchain-postgres==0.0.1
asyncpg==0.28.0
google-generativeai==0.3.2
httpx==0.26.0
python-multipart==0.0.9
pydantic==2.5.3
sqlalchemy>=2.0.29,<3.0.0 